Add files via upload
I have written this Java program with Eclipse. off and on for the past 8 months or so and have changed it a lot as I have learned more about Java and better ways to do what I was going for. I am still cleaning up the code whenever I have free time between work and classes, as well as making it more efficient. 
In the last 3 weeks I have been wanting to use the .stream() function that a coworker told me about and I was finally able to test it out last night. Feel free to add to it and make any suggestions on how I can improve it.